iox-#605 Rename 'relocatable_pointer' dir to 'memory'
