+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/memory/relative_pointer.hpp
@@ -19,7 +19,7 @@
 #define IOX_HOOFS_RELOCATABLE_POINTER_RELATIVE_POINTER_HPP
 
 #include "iceoryx_hoofs/cxx/newtype.hpp"
-#include "iceoryx_hoofs/internal/relocatable_pointer/pointer_repository.hpp"
+#include "iceoryx_hoofs/internal/memory/pointer_repository.hpp"
 
 #include <cstdint>
 #include <iostream>
@@ -211,6 +211,6 @@ PointerRepository<segment_id_underlying_t, UntypedRelativePointer::ptr_t>& getRe
 } // namespace rp
 } // namespace iox
 
-#include "iceoryx_hoofs/internal/relocatable_pointer/relative_pointer.inl"
+#include "iceoryx_hoofs/internal/memory/relative_pointer.inl"
 
 #endif // IOX_HOOFS_RELOCATABLE_POINTER_RELATIVE_POINTER_HPP
