+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/memory/relative_pointer.inl
@@ -18,7 +18,7 @@
 #ifndef IOX_HOOFS_RELOCATABLE_POINTER_RELATIVE_POINTER_INL
 #define IOX_HOOFS_RELOCATABLE_POINTER_RELATIVE_POINTER_INL
 
-#include "iceoryx_hoofs/internal/relocatable_pointer/relative_pointer.hpp"
+#include "iceoryx_hoofs/internal/memory/relative_pointer.hpp"
 
 namespace iox
 {
