+++ iceoryx_posh/include/iceoryx_posh/internal/mepoo/chunk_management.hpp
@@ -18,7 +18,7 @@
 #define IOX_POSH_MEPOO_CHUNK_MANAGEMENT_HPP
 
 #include "iceoryx_hoofs/cxx/helplets.hpp"
-#include "iceoryx_hoofs/internal/relocatable_pointer/relative_pointer.hpp"
+#include "iceoryx_hoofs/internal/memory/relative_pointer.hpp"
 
 #include <atomic>
 #include <cstdint>
